chore: 完成本地后端MVP验收基线

+++ apps/api/app/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from uuid import uuid4
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -38,7 +38,7 @@ class DownloadTask(Base):
     failure_reason: Mapped[str | None] = mapped_column(Text)
     output_filename: Mapped[str | None] = mapped_column(String(255))
     object_key: Mapped[str | None] = mapped_column(Text)
-    object_size: Mapped[int | None] = mapped_column(Integer)
+    object_size: Mapped[int | None] = mapped_column(BigInteger)
     expires_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
     updated_at: Mapped[datetime] = mapped_column(
@@ -61,4 +61,3 @@ class TaskEvent(Base):
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
 
     task: Mapped[DownloadTask] = relationship(back_populates="events")
-
